refactor: rename Btns to authBtns and add Btns component
